refactor: extract `generateWorkerLoader`
